Corrected height of month-name tab
The tab with the name of the month was too high for April, August and September. The letters 'p' and 'g' caused these tabs to be calculated higher than the others. I added a scalebox with a matching y-scale value. It's been at bit trial and error, but I found a size big enough for the text to be considered normal size, small enough to hinder the tab from making itself to high.